+++ packages/mongo/src/builder.ts
@@ -1,5 +1,5 @@
 import { Dict, isNullable, valueMap } from 'cosmokit'
-import { Eval, Field, isComparable, isEvalExpr, Model, Query, Selection, Typed } from 'minato'
+import { Driver, Eval, isComparable, isEvalExpr, Model, Query, Selection, Typed } from 'minato'
 import { Filter, FilterOperators, ObjectId } from 'mongodb'
 import MongoDriver from '.'
 
@@ -70,12 +70,6 @@ function transformFieldQuery(query: Query.FieldQuery, key: string, filters: Filt
   return result
 }
 
-export interface Transformer<S = any, T = any> {
-  types: Field.Type<S>[]
-  dump: (value: S) => T | null
-  load: (value: T, initial?: S) => S | null
-}
-
 export type ExtractUnary<T> = T extends [infer U] ? U : T
 
 export type EvalOperators = {
@@ -95,10 +89,9 @@ export class Builder {
   private refVirtualKeys: Dict<string> = {}
   public aggrDefault: any
 
-  private types: Dict<Transformer> = Object.create(null)
   private evalOperators: EvalOperators
 
-  constructor(private tables: string[], public virtualKey?: string, public recursivePrefix: string = '$') {
+  constructor(private driver: Driver, private tables: string[], public virtualKey?: string, public recursivePrefix: string = '$') {
     this.walkedKeys = []
 
     this.evalOperators = Object.assign(Object.create(null), {
@@ -132,7 +125,7 @@ export class Builder {
       $random: (arg, group) => ({ $rand: {} }),
 
       $cast: (arg, group) => {
-        const converter = this.types[arg[1]]
+        const converter = this.driver.types[arg[1]]
         return converter ? converter.dump(arg[0]) : arg[0]
       },
       $number: (arg, group) => {
@@ -190,18 +183,6 @@ export class Builder {
         })
         return `$${name}`
       },
-    })
-
-    this.define<Buffer, Buffer>({
-      types: ['blob'],
-      dump: value => value,
-      load: (value: any) => value.buffer,
-    })
-
-    this.define<BigInt, string>({
-      types: ['bigint'],
-      dump: value => value ? value.toString() : value as any,
-      load: value => value ? BigInt(value) : value as any,
     })
   }
 
@@ -389,7 +370,7 @@ export class Builder {
   }
 
   protected createSubquery(sel: Selection.Immutable) {
-    const predecessor = new Builder(Object.keys(sel.tables))
+    const predecessor = new Builder(this.driver, Object.keys(sel.tables))
     predecessor.refTables = [...this.refTables, ...this.tables]
     predecessor.refVirtualKeys = this.refVirtualKeys
     return predecessor.select(sel)
@@ -456,15 +437,11 @@ export class Builder {
     return this
   }
 
-  define<S, T>(converter: Transformer<S, T>) {
-    converter.types.forEach(type => this.types[type] = converter)
-  }
-
   dump(model: Model, obj: any): any {
     // obj = model.format(obj)
     const result = {}
     for (const key in obj) {
-      const converter = this.types[model.fields[key]!?.type]
+      const converter = this.driver.types[model.fields[key]!?.type]
       result[key] = converter ? converter.dump(obj[key]) : obj[key]
     }
     // return model.parse(result)
@@ -476,7 +453,7 @@ export class Builder {
   load(model: Model | Typed | Eval.Expr, obj?: any) {
     if (Typed.isTyped(model) || isEvalExpr(model)) {
       const typed = Typed.transform(model)
-      const converter = this.types[typed?.field ?? (typed?.inner ? 'json' : 'raw')]
+      const converter = this.driver.types[typed?.field ?? (typed?.inner ? 'json' : 'raw')]
       return converter ? converter.load(obj) : obj
     }
 
@@ -485,7 +462,7 @@ export class Builder {
     for (const key in obj) {
       if (!(key in model.fields)) continue
       const { type, initial, typed } = model.fields[key]!
-      const converter = typed?.field ? this.types[typed.field] : this.types[type]
+      const converter = typed?.field ? this.driver.types[typed.field] : this.driver.types[type]
       result[key] = converter ? converter.load(obj[key], initial) : obj[key]
     }
     return model.parse(result)
